feat(generation): unify web layout left panel into single scroll column
- 尺寸设置置顶，提示词编辑器随内容自由增高（官网式）
- 移除上下分区与拖拽分隔线（WebPanelSplit 及比例持久化一并删除）
- 钉底控制条压扁并补上批次大小按钮（两种批量生成齐全）
- 参数面板节间距紧凑化；SizeSelector 提取为共享组件
