+++ lib/presentation/providers/layout_state_provider.dart
@@ -17,7 +17,6 @@ class LayoutState {
   final String fixedTagsSidebarViewMode;
   final double fixedTagsNegativeHeight;
   final double webLeftPanelWidth;
-  final double webPromptSectionRatio;
   final bool webLeftPanelExpanded;
   final bool webModelDrawerExpanded;
 
@@ -33,7 +32,6 @@ class LayoutState {
     this.fixedTagsSidebarViewMode = 'list',
     this.fixedTagsNegativeHeight = 180.0,
     this.webLeftPanelWidth = 400.0,
-    this.webPromptSectionRatio = 0.5,
     this.webLeftPanelExpanded = true,
     this.webModelDrawerExpanded = false,
   });
@@ -51,7 +49,6 @@ class LayoutState {
     String? fixedTagsSidebarViewMode,
     double? fixedTagsNegativeHeight,
     double? webLeftPanelWidth,
-    double? webPromptSectionRatio,
     bool? webLeftPanelExpanded,
     bool? webModelDrawerExpanded,
   }) {
@@ -71,8 +68,6 @@ class LayoutState {
       fixedTagsNegativeHeight:
           fixedTagsNegativeHeight ?? this.fixedTagsNegativeHeight,
       webLeftPanelWidth: webLeftPanelWidth ?? this.webLeftPanelWidth,
-      webPromptSectionRatio:
-          webPromptSectionRatio ?? this.webPromptSectionRatio,
       webLeftPanelExpanded: webLeftPanelExpanded ?? this.webLeftPanelExpanded,
       webModelDrawerExpanded:
           webModelDrawerExpanded ?? this.webModelDrawerExpanded,
@@ -100,7 +95,6 @@ class LayoutStateNotifier extends _$LayoutStateNotifier {
       fixedTagsSidebarViewMode: storage.getFixedTagsSidebarViewMode(),
       fixedTagsNegativeHeight: storage.getFixedTagsNegativeHeight(),
       webLeftPanelWidth: storage.getWebLeftPanelWidth(),
-      webPromptSectionRatio: storage.getWebPromptSectionRatio(),
       webLeftPanelExpanded: storage.getWebLeftPanelExpanded(),
       webModelDrawerExpanded: storage.getWebModelDrawerExpanded(),
     );
@@ -217,15 +211,6 @@ class LayoutStateNotifier extends _$LayoutStateNotifier {
 
     final storage = ref.read(localStorageServiceProvider);
     await storage.setWebLeftPanelWidth(clamped);
-  }
-
-  /// 设置官网式布局提示词区高度占比
-  Future<void> setWebPromptSectionRatio(double ratio) async {
-    final clamped = ratio.clamp(0.2, 0.8).toDouble();
-    state = state.copyWith(webPromptSectionRatio: clamped);
-
-    final storage = ref.read(localStorageServiceProvider);
-    await storage.setWebPromptSectionRatio(clamped);
   }
 
   /// 设置官网式布局左栏展开状态
