+++ test/presentation/providers/layout_state_provider_test.dart
@@ -81,7 +81,6 @@ void main() {
       const state = LayoutState();
 
       expect(state.webLeftPanelWidth, 400.0);
-      expect(state.webPromptSectionRatio, 0.5);
       expect(state.webLeftPanelExpanded, isTrue);
       expect(state.webModelDrawerExpanded, isFalse);
     });
@@ -90,14 +89,12 @@ void main() {
       final state = const LayoutState().copyWith(leftPanelWidth: 350.0);
       final updated = state.copyWith(
         webLeftPanelWidth: 480.0,
-        webPromptSectionRatio: 0.6,
         webLeftPanelExpanded: false,
         webModelDrawerExpanded: true,
       );
 
       expect(updated.leftPanelWidth, 350.0);
       expect(updated.webLeftPanelWidth, 480.0);
-      expect(updated.webPromptSectionRatio, 0.6);
       expect(updated.webLeftPanelExpanded, isFalse);
       expect(updated.webModelDrawerExpanded, isTrue);
     });
@@ -107,7 +104,6 @@ void main() {
     test('build 从 storage 读取 web 字段', () {
       final storage = _FakeLayoutStorage()
         ..webWidth = 500.0
-        ..webRatio = 0.35
         ..webExpanded = false
         ..webModelDrawer = true;
       final container = ProviderContainer(
@@ -120,7 +116,6 @@ void main() {
       final state = container.read(layoutStateNotifierProvider);
 
       expect(state.webLeftPanelWidth, 500.0);
-      expect(state.webPromptSectionRatio, 0.35);
       expect(state.webLeftPanelExpanded, isFalse);
       expect(state.webModelDrawerExpanded, isTrue);
     });
@@ -136,20 +131,16 @@ void main() {
 
       final notifier = container.read(layoutStateNotifierProvider.notifier);
       await notifier.setWebLeftPanelWidth(9999.0);
-      await notifier.setWebPromptSectionRatio(0.95);
       await notifier.setWebLeftPanelExpanded(false);
       await notifier.setWebModelDrawerExpanded(true);
 
       expect(storage.webWidth, 560.0);
-      expect(storage.webRatio, 0.8);
       expect(storage.webExpanded, isFalse);
       expect(storage.webModelDrawer, isTrue);
 
       await notifier.setWebLeftPanelWidth(100.0);
-      await notifier.setWebPromptSectionRatio(0.05);
 
       expect(storage.webWidth, 320.0);
-      expect(storage.webRatio, 0.2);
     });
   });
 }
@@ -166,7 +157,6 @@ class _FakeLayoutStorage extends LocalStorageService {
   String viewMode = 'list';
   double negativeHeight = 180.0;
   double webWidth = 400.0;
-  double webRatio = 0.5;
   bool webExpanded = true;
   bool webModelDrawer = false;
 
@@ -226,14 +216,6 @@ class _FakeLayoutStorage extends LocalStorageService {
   @override
   Future<void> setWebLeftPanelWidth(double value) async {
     webWidth = value;
-  }
-
-  @override
-  double getWebPromptSectionRatio() => webRatio;
-
-  @override
-  Future<void> setWebPromptSectionRatio(double value) async {
-    webRatio = value;
   }
 
   @override
